requirements.txt
pyserial
pyyaml
pandas